Fixes edit duplication error (#478)
* Update sitesettings.js

* Update sitesettings.js

* Added the functionality to be able to see and review questions by subject

* added pending questions count and redirection back to question selection page after review

* Update sitesettings.js

* recommit

* recommit 2

* recommit everything

* hides questions that have already been reviewed by the user

* same thing as previous but implemented on all review pages

* added back buttons to review by subject pages

* added functionally to select which question to review

* buttons on all question subpages

button back to respective dashboards

* cleans up internal routing names

* fixes edit duplication error

+++ routes/admin.js
@@ -168,7 +168,7 @@ module.exports = (app, mongo) => {
             { new: true }
         ).then((err, question) => {
             question = question ? question : err.value;
-            if (question && question.reviewers.length > 0) {
+            if (req.isReview && question && question.reviewers.length > 0) {
                 // 2 reviews complete, move to questions collection
                 mongo.db.collection("pendingQuestions").deleteOne({ _id: question._id }, (err, _) => {
                     if (err) {
